requirements.txt
ollama
langchain==0.2.2
langchain-core==0.2.4
langchain_anthropic
langchain-groq
langchain-community 
bs4
pypdf2
chromadb
kaggle
mlflow
scikit-learn
python-dotenv
gradio
langfuse
langserve
sse_starlette


